requirements_txt.txt
streamlit==1.29.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
matplotlib==3.7.2
